feat(trpc): implement role based access

+++ apps/web/server/trpc/router/gameplay.ts
@@ -4,6 +4,7 @@ import { GameplaySchema, GameplayTypes } from '@utils/zod/gameplay';
 import { z } from 'zod';
 import { router, protectedProcedure } from '../trpc';
 import { SegmentSchema } from '@utils/zod/segment';
+import { hasPerms, Perms, Roles } from '@server/utils/hasPerms';
 
 export const gameplayRouter = router({
   get: protectedProcedure
@@ -22,10 +23,15 @@ export const gameplayRouter = router({
       });
 
       // if gameplay not found, or not the user who made it
-      // TODO: need to do role checking
       if (
         gameplay === null ||
-        (gameplay !== null && gameplay.userId !== ctx.session.user.id)
+        !hasPerms({
+          userId: ctx.session.user.id,
+          userRole: Roles.User,
+          itemOwnerId: gameplay.userId,
+          requiredPerms: Perms.isOwner,
+          blacklisted: ctx.session.user.blacklisted,
+        })
       )
         throw new TRPCError({
           code: 'NOT_FOUND',
@@ -44,6 +50,19 @@ export const gameplayRouter = router({
     )
     .output(GameplaySchema)
     .mutation(async ({ input, ctx }) => {
+      // will mostly get thrown if
+      if (
+        !hasPerms({
+          userId: ctx.session.user.id,
+          userRole: Roles.User,
+          requiredPerms: Perms.isOwner,
+          blacklisted: ctx.session.user.blacklisted,
+        })
+      )
+        throw new TRPCError({
+          code: 'UNAUTHORIZED',
+        });
+
       const existingGameplay = await ctx.prisma.footage.findUnique({
         where: {
           youtubeUrl: input.youtubeUrl,
@@ -105,8 +124,20 @@ export const gameplayRouter = router({
     .query(async ({ input, ctx }) => {
       // if no user id provided, use user id from session
       // userId should only be passed by system admins, not avg users
-      // TODO: check roles and prevent users from getting other users
       const userId = input.userId === null ? ctx.session.user.id : input.userId;
+
+      if (
+        !hasPerms({
+          userId: ctx.session.user.id,
+          userRole: Roles.User,
+          itemOwnerId: userId,
+          requiredPerms: Perms.isOwner,
+          blacklisted: ctx.session.user.blacklisted,
+        })
+      )
+        throw new TRPCError({
+          code: 'UNAUTHORIZED',
+        });
 
       const user = await ctx.prisma.user.findUnique({
         where: {
@@ -145,11 +176,23 @@ export const gameplayRouter = router({
       });
 
       // if gameplay not found, or not the user who made it
-      // TODO: need to do role checking
-      if (gameplay === null || gameplay.userId !== ctx.session.user.id)
+      if (gameplay === null)
         throw new TRPCError({
           code: 'NOT_FOUND',
           message: 'Could not find that requested gameplay',
+        });
+
+      if (
+        !hasPerms({
+          userId: ctx.session.user.id,
+          userRole: Roles.User,
+          itemOwnerId: gameplay.userId,
+          requiredPerms: Perms.isOwner,
+          blacklisted: ctx.session.user.blacklisted,
+        })
+      )
+        throw new TRPCError({
+          code: 'UNAUTHORIZED',
         });
 
       return gameplay.clips;
@@ -165,14 +208,33 @@ export const gameplayRouter = router({
     )
     .output(GameplaySchema)
     .mutation(async ({ input, ctx }) => {
-      // TODO: need check if user making request is the
-      // user who owns the gameplay
-
-      // TODO: need to prevent users from modifying isAnalyzed
-      // need to do a role check here
-
       try {
-        const gameplay = await ctx.prisma.footage.update({
+        const gameplay = await ctx.prisma.footage.findUniqueOrThrow({
+          where: {
+            id: input.gameplayId,
+          },
+        });
+
+        // if modifying isAnalyzed, require mod role
+        const requiredPerms =
+          gameplay.isAnalyzed === input.isAnalyzed
+            ? Perms.isOwner
+            : Perms.roleMod;
+
+        if (
+          !hasPerms({
+            userId: ctx.session.user.id,
+            userRole: Roles.User,
+            itemOwnerId: gameplay.userId,
+            requiredPerms,
+            blacklisted: ctx.session.user.blacklisted,
+          })
+        )
+          throw new TRPCError({
+            code: 'UNAUTHORIZED',
+          });
+
+        const updatedGameplay = await ctx.prisma.footage.update({
           where: {
             id: input.gameplayId,
           },
@@ -182,7 +244,7 @@ export const gameplayRouter = router({
           },
         });
 
-        return gameplay;
+        return updatedGameplay;
       } catch (error) {
         // throws RecordNotFound if record not found to update
         // but can't import for some reason
@@ -203,10 +265,26 @@ export const gameplayRouter = router({
       }),
     )
     .mutation(async ({ input, ctx }) => {
-      // TODO: need check if user making request is the
-      // user who owns the gameplay
-
       try {
+        const gameplay = await ctx.prisma.footage.findUniqueOrThrow({
+          where: {
+            id: input.gameplayId,
+          },
+        });
+
+        if (
+          !hasPerms({
+            userId: ctx.session.user.id,
+            userRole: Roles.User,
+            itemOwnerId: gameplay.userId,
+            requiredPerms: Perms.isOwner,
+            blacklisted: ctx.session.user.blacklisted,
+          })
+        )
+          throw new TRPCError({
+            code: 'UNAUTHORIZED',
+          });
+
         await ctx.prisma.footage.delete({
           where: {
             id: input.gameplayId,
